requirements.txt
requests
beautifulsoup4
pandas
pyarrow
python-dotenv
lxml
streamlit
openai
